# Project examples/Go-TS/foxgres/cmd/foxgres/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ориокс здорового человека",
        "title": "foxgres API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/auth": {
            "get": {
                "description": "Поиск пользователя в таблице users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин",
                        "name": "login",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Пароль",
                        "name": "password",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AccessLevel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ResponseErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ResponseErr"
                        }
                    }
                }
            }
        },
        "/marks": {
            "get": {
                "description": "Получение оценок студента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "get student marks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор студента",
                        "name": "student_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Marks"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ResponseErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AccessLevel": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer"
                }
            }
        },
        "Marks": {
            "type": "object",
            "properties": {
                "marks": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "student": {
                    "$ref": "#/definitions/Student"
                }
            }
        },
        "ResponseErr": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Student": {
            "type": "object",
            "properties": {
                "group_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}